requirements.txt
akshare==1.16.81
backtrader==1.9.78.123
loguru==0.7.3
pandas==2.2.3
pre-commit==4.2.0
pydantic==2.11.3
pyecharts==2.0.8
PyYAML==6.0.2
streamlit==1.44.1
streamlit_echarts==0.4.0
